very_basic_fastapi_graphql/requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn
sqlalchemy 
psycopg2-binary
python-dotenv